frontend/requirements.txt
streamlit
sqlalchemy
python-dotenv